fix: Resolve compilation errors in knhk-consensus, knhk-closed-loop, and knhk-kernel
- knhk-consensus/hotstuff.rs: Make mutable variables for vote/prepare count operations
- knhk-consensus/network.rs: Fix borrow of moved value by cloning msg.source before move
- knhk-closed-loop: Fix import issues, adapt to actual DetectedPattern structure, fix LLM integration
- knhk-kernel/executor.rs: Cast u32 tick_budget to u64 for TickBudget::with_budget()
- knhk-kernel/guard.rs: Fix lifetime issue by cloning guard before moving into closure
- knhk-kernel/hot_path.rs: Fix Receipt cloning with into_iter().map()
- knhk-kernel/receipt.rs: Add Clone derive to Receipt struct for cloning support

All modified modules now compile successfully without errors.

+++ rust/knhk-kernel/src/executor.rs
@@ -170,7 +170,7 @@ impl Executor {
     #[inline(always)]
     pub fn execute(&self, task: &Task) -> Receipt {
         let timer = HotPathTimer::start();
-        let mut budget = TickBudget::with_budget(task.tick_budget);
+        let mut budget = TickBudget::with_budget(task.tick_budget as u64);
 
         // Start receipt
         let mut receipt = ReceiptBuilder::new(task.pattern_id, task.task_id)
